Mostly Building Stuff
- Buildings make noises when placed and picked up
- Buildings are now drawn relative to where they were picked up in relation to the mouse
- Buildings can be Upgraded
- Maps refer to the BuildingID they represent
- Crafting is now done in seconds
- All craft recipes are in the form of a dictionary
- Cleaned up some of the build mode right click handling
- Player stats now represented by doubles
- Added method to Calendar to determine elapsed time from bedtime to next dawn

+++ RiverHollowGame/GUIComponents/GUIObjects/GUIWindows/GUIStatDisplay.cs
@@ -12,11 +12,11 @@ namespace RiverHollow.GUIComponents.GUIObjects.GUIWindows
         GUIImage _gFill;
         GUIText _gText;
 
-        int _iMax;
-        int _iCurr;
+        double _iMax;
+        double _iCurr;
         const int EDGE = 4;
 
-        public delegate void DelegateRetrieveValues(ref int curr, ref int max);
+        public delegate void DelegateRetrieveValues(ref double curr, ref double max);
         DelegateRetrieveValues _delAction;
 
         public GUIStatDisplay(DelegateRetrieveValues del, Color c, int width = 200) : base(DisplayWin, width, GameManager.ScaledTileSize/2)
